refactor: separate metadata orchestration from label and annotation helpers
Move ensureInProgressMetadata to its own metadata.go file since it
orchestrates both labels and annotations. Keep labels.go for label-only
helpers and annotations.go for annotation-only helpers.

Also add shared-routes documentation page to mkdocs.yml navigation.
