+++ mkdocs.yml
@@ -22,6 +22,7 @@ nav:
 - Features: 
   - Advanced Deployments: features/advanced-deployments.md
   - Multiple Routes: features/multiple-routes.md
+  - Shared Routes: features/shared-routes.md
   - Header Based Routing: features/header-based-routing.md    
   - TCP Routing: features/tcp.md
   - TLS Routing: features/tls.md
